[Mail] Preserve 64-bit money and COD
Keep unsigned mail values intact through storage and API boundaries. Add overflow-safe policy predicates and mutation-backed coverage while leaving dormant mail routes unregistered.

+++ src/game/Object/ObjectMgr.cpp
@@ -693,7 +693,7 @@ void ObjectMgr::ReturnOrDeleteOldMails(bool serverUp)
         bool has_items = fields[4].GetBool();
         m->expire_time = (time_t)fields[5].GetUInt64();
         m->deliver_time = 0;
-        m->COD = fields[6].GetUInt32();
+        m->COD = fields[6].GetUInt64();
         m->checked = fields[7].GetUInt32();
         m->mailTemplateId = fields[8].GetInt16();
